Restore full app functionality with CSRF fix
